Payload ve logda iyileştirmeler. - Loglamayı açma kapama için config - Payload Uzunluğu 1 byte yetmediği için 2 byte'a yükseltildi.

+++ rfd/packets/headers/AckPacket.h
@@ -10,9 +10,9 @@ public:
     AckPacket(uint64_t sender,uint64_t receiver, uint8_t type) 
         : Packet(PACKET_ACK, sender), ack_type(type), receiver(receiver) {}
 
-    uint8_t getPayloadSize() const override { return Packet::getPayloadSize() + sizeof(ack_type) + sizeof(receiver); }
+    uint16_t getPayloadSize() const override { return Packet::getPayloadSize() + sizeof(ack_type) + sizeof(receiver); }
     void serialize(uint8_t* buffer) const override;
-    bool deserialize(const uint8_t* buffer, uint8_t length) override;
+    bool deserialize(const uint8_t* buffer, uint16_t length) override;
     void handle(Node& node_instance) override;
 
     uint8_t getAckType() const { return ack_type; }
