+++ rfd/packets/headers/HelloPacket.h
@@ -13,9 +13,9 @@ public:
     HelloPacket(uint64_t sender, uint64_t pub_key, uint64_t mod) 
         : Packet(PACKET_HELLO, sender), public_key(pub_key), modulus(mod) {}
 
-    uint8_t getPayloadSize() const override { return Packet::getPayloadSize() + sizeof(public_key) + sizeof(modulus); }
+    uint16_t getPayloadSize() const override { return Packet::getPayloadSize() + sizeof(public_key) + sizeof(modulus); }
     void serialize(uint8_t* buffer) const override;
-    bool deserialize(const uint8_t* buffer, uint8_t length) override;
+    bool deserialize(const uint8_t* buffer, uint16_t length) override;
     void handle(Node& node_instance) override;
 
     // Getters
