+++ rfd/packets/headers/KeyPacket.h
@@ -11,11 +11,11 @@ public:
     KeyPacket(uint64_t sender, uint64_t target, uint64_t pub_key, uint64_t mod)
         : Packet(PACKET_KEY, sender), target_id(target), public_key(pub_key), modulus(mod) {}
 
-    uint8_t getPayloadSize() const override { 
+    uint16_t getPayloadSize() const override { 
         return Packet::getPayloadSize() + sizeof(target_id) + sizeof(public_key) + sizeof(modulus);
     }
     void serialize(uint8_t* buffer) const override;
-    bool deserialize(const uint8_t* buffer, uint8_t length) override;
+    bool deserialize(const uint8_t* buffer, uint16_t length) override;
     void handle(Node& node_instance) override;
 
     uint64_t getTargetId() const { return target_id; }
