+++ rfd/packets/headers/PingPacket.h
@@ -9,11 +9,11 @@ public:
     PingPacket(uint64_t sender, uint64_t target, uint32_t ts) 
         : Packet(PACKET_PING, sender), target_id(target) {}
 
-    uint8_t getPayloadSize() const override { 
+    uint16_t getPayloadSize() const override { 
         return Packet::getPayloadSize() + sizeof(target_id); 
     }
     void serialize(uint8_t* buffer) const override;
-    bool deserialize(const uint8_t* buffer, uint8_t length) override;
+    bool deserialize(const uint8_t* buffer, uint16_t length) override;
     void handle(Node& node_instance) override;
 
     uint64_t getTargetId() const { return target_id; }
