+++ rfd/packets/src/PingPacket.cpp
@@ -11,7 +11,7 @@ void PingPacket::serialize(uint8_t* buffer) const {
     offset += sizeof(target_id);
 }
 
-bool PingPacket::deserialize(const uint8_t* buffer, uint8_t length) {
+bool PingPacket::deserialize(const uint8_t* buffer, uint16_t length) {
     if (length < getPayloadSize()) return false;
     Packet::deserialize(buffer, length);
     size_t offset = Packet::getPayloadSize();
